Move chacha20 functionality to a new file
