+++ lib/crypto/c_src/Makefile.in
@@ -74,6 +74,7 @@ LIBDIR = $(PRIVDIR)/lib/$(TARGET)
 CRYPTO_OBJS = $(OBJDIR)/crypto$(TYPEMARKER).o \
 	$(OBJDIR)/atoms$(TYPEMARKER).o \
 	$(OBJDIR)/bn$(TYPEMARKER).o \
+	$(OBJDIR)/chacha20$(TYPEMARKER).o \
 	$(OBJDIR)/cipher$(TYPEMARKER).o \
 	$(OBJDIR)/cmac$(TYPEMARKER).o \
 	$(OBJDIR)/dh$(TYPEMARKER).o \
